fix: 명시적 타입 변환으로 오버플로우 문제 방지

+++ src/main/java/domain/WinningCase.java
@@ -38,7 +38,7 @@ public enum WinningCase {
   }
 
   public long calculateEarnMoney(int winningCaseCount) {
-    return winningCaseCount * winningMoney;
+    return  (long) winningCaseCount * winningMoney;
   }
 
   public int getWinningMoney() {
